+++ initialize.h
@@ -3,4 +3,4 @@
 #include "sugar.h"
 #include "agent.h"
 
-void initialize_agent(Simulation *, Agent *, int, int, int, int, int, int, int);
+void initialize_agent(Simulation *, Agent *, int, int, int, int, int, float, float);
